Cleanup of Hotel Listing Page Code: - Modularized all code (made each section its own component). - Improved and optimized ImageModal to use the Modal comp I already made. - Changed ModalRoute to Modal for more to fit the purpose of the Modal more clearly.
