+++ types/rapid-hotels-api/CompleteHotelInformation.ts
@@ -1,0 +1,7 @@
+import { APIHotelDetailsJSONFormatted } from "@/app/api/hotels/details/route";
+import { APIHotelRoomOffersJSONFormatted } from "@/app/api/hotels/search/rooms/route";
+
+export type CompleteHotelInfo = {
+    hotelDetails: APIHotelDetailsJSONFormatted | null;
+    hotelRooms: APIHotelRoomOffersJSONFormatted | null;
+} | null;
